Remove byte calculations from instructions
