+++ instructions.cc
@@ -2,6 +2,7 @@
 #include <bitset>
 #include <vector>
 #include <map>
+#include "bytes.h"
 
 //Instruction Class
 std::string Instruction::to_string()
@@ -108,36 +109,7 @@ Instruction *moveHighToLow(Registers low, Registers high)
         low);
 }
 
-std::string getVal5Bits(int val)
-{
-    std::bitset<5> bs(val);
-    return bs.to_string();
-}
 
-std::string getVal3Bits(int val)
-{
-    std::bitset<3> bs(val);
-    return bs.to_string();
-}
-
-std::string getVal4Bits(int val)
-{
-    std::bitset<4> bs(val);
-    return bs.to_string();
-}
-
-std::string getVal8Bits(int val)
-{
-    std::bitset<8> bs(val);
-    return bs.to_string();
-}
-
-std::string getVal8BitsSignal(int val)
-{
-    std::bitset<8> bs(val);
-    // TODO : Add support to signal
-    return bs.to_string();
-}
 
 void Instruction::setlabel(std::string newLabel)
 {
